# admin-be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Backend",
        "description": "Admin Backend Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://admin-apc.dotbeit.com",
            "description": "Admin Backend"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "User Login",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::authenticate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "required|string|email|max:255",
                                        "type": "string",
                                        "example": "user@user.com"
                                    },
                                    "password": {
                                        "description": "required|string|min:6",
                                        "type": "string",
                                        "example": "123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token:": {
                                            "description": "The Token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid_credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error:These credentials do not match our records.": {
                                            "description": "These credentials do not match our records.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Validator.": {
                                            "description": "Validator",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/events": {
            "get": {
                "tags": [
                    "Event Pages"
                ],
                "operationId": "events",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['event']"
                    }
                }
            }
        },
        "/api/event/{slug}": {
            "get": {
                "tags": [
                    "Event Pages"
                ],
                "operationId": "event",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "cardiac-diseases-event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['event']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
        "/api/event-user": {
            "post": {
                "tags": [
                    "Event Pages"
                ],
                "summary": "User Events",
                "operationId": "App\\Http\\Controllers\\Api\\EventController::eventUserAction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id"
                                ],
                                "properties": {
                                    "event_id": {
                                        "description": "required|numeric",
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "event successfully removed form your favorite list :(",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "event successfully removed",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "event successfully added to your favorite list :)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success:": {
                                            "description": "event successfully added ",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user_not_found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_not_found": {
                                            "description": "user_not_found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Validator.": {
                                            "description": "Validator",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/category/{slug}": {
            "get": {
                "tags": [
                    "Post Pages"
                ],
                "operationId": "category",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "category-1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['Category']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
...
        "/api/service/{slug}": {
            "get": {
                "tags": [
                    "Service Pages"
                ],
                "operationId": "service",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "وحدة-قسطرة-القلب"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['service']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}